Minor refactors

+++ src/constants.rs
@@ -9,8 +9,8 @@ pub(crate) const THREADED_BITS: usize = 3; // num threads = 2^THREADED_BITS
 pub(crate) const CHANNEL_SIZE: usize = 100 * 1024;
 
 // Calculated
-pub(crate) const NUM_THREADS: LineMaskType = 2u32.pow(THREADED_BITS as u32);
+pub(crate) const NUM_THREADS: LineMaskType = 2usize.pow(THREADED_BITS as u32);
 pub(crate) const NUM_HASH_BYTES: usize = (DESIRED_HEX_MATCHES + 1) / 2;
 
 // Fixed
-pub(crate) type LineMaskType = u32;
+pub(crate) type LineMaskType = usize;
